requirements.txt
fastapi==0.87.0
pydantic==1.8.2
python-dotenv==0.21.0
redis==4.3.4
transformers==4.18.0
uvicorn==0.17.6
torch==1.13.0